feat: 필터 디테일 뷰에서 작가 프로필 탭시 작가 디테일로 이동하도록 네비게이션 패스 설정과 구현 필터 디테일뷰 사진 정보 포멧 수정

+++ ShutterLink/Model/Network/DTOs/FilterDetail.swift
@@ -31,7 +31,7 @@ struct FilterDetailResponse: Decodable {
 struct PhotoMetadata: Decodable {
     let camera: String
     let lens_info: String
-    let focal_length: Int
+    let focal_length: Float
     let aperture: Double
     let iso: Int
     let shutter_speed: String
